+++ common/start_build.sh
@@ -11,42 +11,20 @@ setup_dist_dirs "$STEAM_APP_ID_LIST"
 
 echo "APP_IDS=$STEAM_APP_ID_LIST" >> $GITHUB_ENV
 
-if [ ! -z "${CUSTOM_CONTAINER}" ]; then
-   setup_custom_container
-   
-    if [ ! -z "${GCC_10}" ]; then
-        echo "Installing gcc 10"
-        install_gcc_10_ubuntu_1804
-    fi
+git config --global user.email "actions@github.com"
+git config --global user.name "GitHub Action"
 
-    if [ ! -z "${GCC_9}" ]; then
-        echo "Installing gcc 9"
-        install_gcc_9_ubuntu_1804
-    fi
-else
-    cp -rfv ../../common/ubuntu12sources.list /etc/apt/sources.list
-
-    if [ ! -z "${GCC_9}" ]; then
-        echo "Installing gcc 9"
-        install_gcc_9
-    fi
-
-    if [ ! -z "${GCC_6}" ]; then
-        echo "Installing gcc 6"
-        install_gcc_6
-    fi
-
-    if [ ! -z "${LATEST_GIT}" ]; then
-        echo "Installing latest git"
-        install_latest_git
-    fi
-fi
-
-if [ ! -z "${COMMON_QT5}" ]; then
-   use_common_qt5
-fi
+install_latest_cmake
 
 gcc --version
+
+if [ ! -z "${LIBRARIES}" ]; then
+    echo "Found libraries to build: $LIBRARIES"
+    pushd ../../libraries
+    source start_library_build.sh
+    start_library_build "$LIBRARIES"
+    popd
+fi
 
 source ./build.sh