+++ engines/augustus/build.sh
@@ -18,5 +18,4 @@ popd
 cp -rfv "source/build/augustus" "$diststart/517790/dist/"
 cp -rfv "source/assets" "$diststart/517790/dist/"
 cp -rfv "source/res/maps" "$diststart/517790/dist/"
-
 cp -rfv assets/* "$diststart/517790/dist/"