+++ engines/rottexpr/env.sh
@@ -2,5 +2,3 @@
 
 export STEAM_APP_ID_LIST="238050"
 export LICENSE_PATH="./source/COPYING"
-export ADDITIONAL_LICENSES="./SDL/COPYING.txt ./SDL_mixer/COPYING.txt"
-export CUSTOM_CONTAINER="ubuntu:18.04"