+++ libraries/boostheaders/env.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+export LIBRARY_LICENSES="./boost/LICENSE_1_0.txt"