+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 32.0 (2021-09-16)
+
+* Add support for building against steam runtime
+* Added support for common library build scripts, so engines that share libraries do not have to each define the build scripts for them. Now can set ```export LIBRARIES="fluidsynth zmusic"``` as an example, which will clone, build, and copy the license and library files to the artifact.
+
 ### 30.4 (2021-09-14)
 
 * dhewm3 - Add demo support